Mobile design completed
