+++ css/style.css
@@ -1,0 +1,101 @@
+*{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+}
+/*Header Sections*/
+header {
+    margin: 1em;
+}
+
+nav {
+    display: flex;
+    justify-content: space-between;
+}
+
+nav .bx {
+    font-size: 1.8rem;
+}
+
+nav img {
+    width: 10em;
+}
+
+/*Mobile Menus*/
+.hidden {
+    display: none;
+}
+.mobile-menu ul {
+    margin: 1em 0;
+    list-style: none;
+    justify-content: space-between;
+}
+
+.mobile-menu ul li {
+    padding: 1em;
+    justify-content: center;
+    transition: all ease 0.5s;
+}
+
+/*Hero BUtton section */
+
+.hero-btn-section {
+    margin: 1.5em 1em;
+    text-align: center;
+}
+
+.hero-btn-section button {
+    padding: 0.5em;
+    width: 10em;
+    font-size: 1rem;
+    border-radius: 40px;
+    border: 3px solid #FFA724;
+    background-color: transparent;
+    transition: all ease 0.5s;
+}
+
+.account-spendings-section {
+    margin: 1em;
+}
+
+.account-section ul {
+    list-style: none;
+}
+
+.account-section ul li {
+    display: flex;
+    justify-content: space-between;
+    margin: 1em 0;
+    padding: 1.5em;
+    border: 1px solid #FFA724;
+    border-radius: 5px;
+    transition: all ease 0.5s;
+}
+
+.account-section ul li:hover, 
+.account-section ul li:active, 
+.active, 
+.hero-btn-section button:hover, 
+.mobile-menu ul li:hover,
+.mobile-menu ul li:active
+{
+    background-color: #FFA724;
+    cursor: pointer;
+}
+
+.spending-section ul {
+    list-style: none;
+}
+
+.spending-section ul li {
+    display: flex;
+    justify-content: space-between;
+    margin: 1em 0;
+    padding: 1em;
+    border: 1px solid #FFA724;
+    border-radius: 5px;
+    background-color: #FFA724;
+}
+
+
