--- hehe
@@ -1,32 +1,0 @@
-
-// const clickBtn = document.querySelector(".account-details");
-
-
-
-// clickBtn.addEventListener('click', function(e) {
-//     if (e.target.tagName === 'LI') {
-//         const liId = e.target.id;
-//         switch (liId) {
-//             case 'mainAccount':
-//                 mainAccount();
-//                 break;
-
-//             case 'expenses':
-//                 expenses()
-//                 break;
-
-//             case 'saving':
-//                 saving()
-//                 break;
-
-//             default:
-//                 break;
-//         }
-//     }
-// });
-
-
-
-
-
-//  console.log(accounts.title)
